Fixed a regression brought in #1123
When introducing the `inputs` config key, the default behavior when none
of the `inputs`, `preamble`, `index` and `postamble` key is given has
been changed to use an empty list instead of the `DEFAULT_INPUTS` new
default `const`.

If one of the 3 legacy fields is used, we will use them and warn the user
about its deprecated state. Otherwise, use the default `inputs` value.

Signed-off-by: Oror <Lugrim@users.noreply.github.com>